finish feature implement api ch5

+++ app/src/main/java/com/catnip/ninsfood_binarch3/presentation/feature/detailproduct/DetailProductViewModel.kt
@@ -21,8 +21,8 @@ class DetailProductViewModel(
         postValue(0)
     }
 
-    val priceLiveData = MutableLiveData<Double>().apply {
-        postValue(0.0)
+    val priceLiveData = MutableLiveData<Int>().apply {
+        postValue(0)
     }
 
     private val _addToCartResult = MutableLiveData<ResultWrapper<Boolean>>()
@@ -39,14 +39,14 @@ class DetailProductViewModel(
     fun add() {
         val count = (productCountLiveData.value ?: 0) + 1
         productCountLiveData.postValue(count)
-        priceLiveData.postValue(product?.price?.times(count) ?: 0.0)
+        priceLiveData.postValue(product?.price?.times(count) ?: 0)
     }
 
     fun minus() {
         if((productCountLiveData.value ?: 0) > 0){
             val count = (productCountLiveData.value ?: 0) -1
             productCountLiveData.postValue(count)
-            priceLiveData.postValue(product?.price?.times(count) ?: 0.0)
+            priceLiveData.postValue(product?.price?.times(count) ?: 0)
         }
     }
